test qltk

+++ src/main/resources/static/admin/authorize/authorize-ctrl.js
@@ -1,24 +1,28 @@
-app.controller("ctrl", function($scope, $http){
+app.controller("ctrl", function($scope, $http) {
 	$http.get("/rest/authorities").then(resp => {
 		$scope.db = resp.data;
 	})
-	
-	$scope.index_of = function(username, role){
+
+	$scope.index_of = function(username, role) {
 		return $scope.db.authorities
 			.findIndex(a => a.account.username == username && a.role.id == role);
 	}
-	
-	$scope.update = function(username, role){
+
+	$scope.update = function(username, role) {
 		var index = $scope.index_of(username, role);
-		if(index >= 0){
+		alert(index)
+		if (index >= 0) {
 			var id = $scope.db.authorities[index].id;
-			$http.delete(`/rest/authorities/${id}`).then(resp =>{
-				$scope.db.authorities.splice(index, 1);
-			})
-		}else{
+			var username = $scope.db.authorities[index].account.username;
+			if (username == "hsang") {
+				$http.delete(`/rest/authorities/${id}`).then(resp => {
+					$scope.db.authorities.splice(index, 1);
+				})
+			}
+		} else {
 			var authority = {
-				account:{username: username},
-				role:{id: role}
+				account: { username: username },
+				role: { id: role }
 			};
 			$http.post('/rest/authorities', authority).then(resp => {
 				$scope.db.authorities.push(resp.data);
